Convert the client app to use the cobra command-line framework.
