+++ example/client/cmd/quota.go
@@ -1,0 +1,34 @@
+// Copyright 2016 Google Inc.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package cmd
+
+import (
+	"fmt"
+
+	"github.com/spf13/cobra"
+)
+
+// quotaCmd represents the quota command
+var quotaCmd = &cobra.Command{
+	Use:   "quota",
+	Short: "Invokes the mixer's Quota API.",
+	Run: func(cmd *cobra.Command, args []string) {
+		fmt.Println("Not implemented yet")
+	},
+}
+
+func init() {
+	RootCmd.AddCommand(quotaCmd)
+}
